requirements.txt
black==24.2.0
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pymongo==4.6.2
celery==5.3.6
django-treebeard==4.7.1
